# backend/openapi.json
{
  "components": {
    "schemas": {
      "AccessMode": {
        "enum": [
          "ssh",
          "agent"
        ],
        "title": "AccessMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ServerRegisterRequest": {
        "description": "纳管 API 入参(§3.2)。\n\n与 ServerCreate 的区别:SSH 模式收**私钥明文**（ssh_private_key）而非\ncredential_id——私钥由 API 层存入保险箱换取 credential_id，绝不落业务表（§13）。",
        "properties": {
          "access_mode": {
            "$ref": "#/components/schemas/AccessMode"
          },
          "agent_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "auth_type": {
            "default": "key",
            "title": "Auth Type",
            "type": "string"
          },
          "environment": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Environment",
            "type": "string"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "labels": {
            "additionalProperties": true,
            "title": "Labels",
            "type": "object"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "ssh_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh Password"
          },
          "ssh_port": {
            "default": 22,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Ssh Port",
            "type": "integer"
          },
          "ssh_private_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh Private Key"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "name",
          "host",
          "access_mode",
          "environment"
        ],
        "title": "ServerRegisterRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "一脉 Axon 控制面",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/servers": {
      "get": {
        "operationId": "list_servers_api_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Servers Api Servers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Servers",
        "tags": [
          "servers"
        ]
      },
      "post": {
        "operationId": "register_server_api_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Server Api Servers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Server",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{server_id}": {
      "delete": {
        "operationId": "delete_server_api_servers__server_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Server Api Servers  Server Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Server",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{server_id}/install-agent": {
      "post": {
        "description": "对 SSH 纳管的服务器经 SSH 下发安装 axon-agent(需求4)。\n\n落一条 agent_install task,后台经 SSHExecutor 跑安装脚本(二进制从控制面下载\n端点拉取,离线分发),前端轮询 task 终态。非 SSH 服务器由编排层落 task.failed。",
        "operationId": "install_agent_api_servers__server_id__install_agent_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Install Agent Api Servers  Server Id  Install Agent Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Install Agent",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{server_id}/test-connection": {
      "post": {
        "operationId": "test_connection_api_servers__server_id__test_connection_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Connection Api Servers  Server Id  Test Connection Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Connection",
        "tags": [
          "servers"
        ]
      }
    },
...
  }
}